// client/src/domain/sdcForm/SDCForm.ts
export interface SDCFormModel {
  metadata: Partial<SDCFormMetaData>;
  SDCFormId: string;
  title: string;
  version: string;
  diagnosticProcedureId: string;
  body: SDCSection;
  footer: SDCFormFooter;
}
// ...
const SDCFormRecord = immutableRecord<SDCFormModel>({
  metadata: {},
  SDCFormId: '',
  title: '',
  version: '',
  diagnosticProcedureId: '',
  body: {
    minRepetitions: 1,
    maxRepetitions: 1,
  },
  footer: {},
});

export class SDCForm extends SDCFormRecord {}